# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FSM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobListResponse" }
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResponse" }
  /jobs/{id}:
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResponse" }
    delete:
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceListResponse" }
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /trucks:
    get:
      operationId: listTrucks
      tags: [fleet]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TruckListResponse" }
    post:
      operationId: createTruck
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TruckWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TruckResponse" }
  /trucks/{id}:
    patch:
      operationId: updateTruck
      tags: [fleet]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TruckWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TruckResponse" }
    delete:
      operationId: deleteTruck
      tags: [fleet]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /equipment:
    get:
      operationId: listEquipment
      tags: [fleet]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentListResponse" }
    post:
      operationId: createEquipment
      tags: [fleet]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipmentWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentResponse" }
  /equipment/{id}:
    patch:
      operationId: updateEquipment
      tags: [fleet]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipmentWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentResponse" }
    delete:
      operationId: deleteEquipment
      tags: [fleet]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /maintenance-logs:
    get:
      operationId: listMaintenanceLogs
      tags: [maintenance]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceLogListResponse" }
    post:
      operationId: createMaintenanceLog
      tags: [maintenance]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaintenanceLogWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceLogResponse" }
  /maintenance-logs/{id}:
    patch:
      operationId: updateMaintenanceLog
      tags: [maintenance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaintenanceLogWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceLogResponse" }
    delete:
      operationId: deleteMaintenanceLog
      tags: [maintenance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
...
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Job:
      type: object
      required: [id, propertyId, status, totalCents, createdAt]
      properties:
        id: { type: integer }
        propertyId: { type: integer }
        crewId: { type: integer, nullable: true }
        status: { type: string }
        scheduledFor: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        totalCents: { type: integer }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    JobWriteBody:
      type: object
      required: [propertyId, status, totalCents]
      properties:
        propertyId: { type: integer }
        crewId: { type: integer, nullable: true }
        status: { type: string }
        scheduledFor: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        totalCents: { type: integer }
        notes: { type: string, nullable: true }
    JobResponse:
      type: object
      required: [job]
      properties:
        job: { $ref: "#/components/schemas/Job" }
    JobListResponse:
      type: object
      required: [jobs]
      properties:
        jobs:
          type: array
          items: { $ref: "#/components/schemas/Job" }
...
    Invoice:
      type: object
      required: [id, customerId, status, totalCents]
      properties:
        id: { type: integer }
        jobId: { type: integer, nullable: true }
        customerId: { type: integer }
        status: { type: string }
        totalCents: { type: integer }
        issuedAt: { type: string, format: date-time, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
    InvoiceWriteBody:
      type: object
      required: [customerId, status, totalCents]
      properties:
        jobId: { type: integer, nullable: true }
        customerId: { type: integer }
        status: { type: string }
        totalCents: { type: integer }
        issuedAt: { type: string, format: date-time, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
    InvoiceResponse:
      type: object
      required: [invoice]
      properties:
        invoice: { $ref: "#/components/schemas/Invoice" }
    InvoiceListResponse:
      type: object
      required: [invoices]
      properties:
        invoices:
          type: array
          items: { $ref: "#/components/schemas/Invoice" }
    Truck:
      type: object
      required: [id, name, status]
      properties:
        id: { type: integer }
        name: { type: string }
        vin: { type: string, nullable: true }
        plate: { type: string, nullable: true }
        status: { type: string }
        assignedCrewId: { type: integer, nullable: true }
    TruckWriteBody:
      type: object
      required: [name, status]
      properties:
        name: { type: string }
        vin: { type: string, nullable: true }
        plate: { type: string, nullable: true }
        status: { type: string }
        assignedCrewId: { type: integer, nullable: true }
    TruckResponse:
      type: object
      required: [truck]
      properties:
        truck: { $ref: "#/components/schemas/Truck" }
    TruckListResponse:
      type: object
      required: [trucks]
      properties:
        trucks:
          type: array
          items: { $ref: "#/components/schemas/Truck" }
    Equipment:
      type: object
      required: [id, name, type, status]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        serial: { type: string, nullable: true }
        status: { type: string }
        assignedTruckId: { type: integer, nullable: true }
    EquipmentWriteBody:
      type: object
      required: [name, type, status]
      properties:
        name: { type: string }
        type: { type: string }
        serial: { type: string, nullable: true }
        status: { type: string }
        assignedTruckId: { type: integer, nullable: true }
    EquipmentResponse:
      type: object
      required: [equipment]
      properties:
        equipment: { $ref: "#/components/schemas/Equipment" }
    EquipmentListResponse:
      type: object
      required: [equipment]
      properties:
        equipment:
          type: array
          items: { $ref: "#/components/schemas/Equipment" }
    MaintenanceLog:
      type: object
      required: [id, kind, description, performedAt, costCents]
      properties:
        id: { type: integer }
        truckId: { type: integer, nullable: true }
        equipmentId: { type: integer, nullable: true }
        kind: { type: string }
        description: { type: string }
        performedByUserId: { type: integer, nullable: true }
        performedAt: { type: string, format: date-time }
        costCents: { type: integer }
    MaintenanceLogWriteBody:
      type: object
      required: [kind, description, costCents]
      properties:
        truckId: { type: integer, nullable: true }
        equipmentId: { type: integer, nullable: true }
        kind: { type: string }
        description: { type: string }
        performedByUserId: { type: integer, nullable: true }
        performedAt: { type: string, format: date-time, nullable: true }
        costCents: { type: integer }
    MaintenanceLogResponse:
      type: object
      required: [log]
      properties:
        log: { $ref: "#/components/schemas/MaintenanceLog" }
    MaintenanceLogListResponse:
      type: object
      required: [logs]
      properties:
        logs:
          type: array
          items: { $ref: "#/components/schemas/MaintenanceLog" }
...