# lib/api-zod/src/generated/types/equipmentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Equipment } from "./equipment";

export interface EquipmentResponse {
  equipment: Equipment;
}