# lib/api-zod/src/generated/types/equipmentWriteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentWriteBody {
  name: string;
  type: string;
  serial?: string | null;
  status: string;
  assignedTruckId?: number | null;
}