# lib/api-zod/src/generated/types/invoiceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";

export interface InvoiceResponse {
  invoice: Invoice;
}